offline: enable localhost PWA testing.
`ng serve` seems to compile in-memory and not keep ngsw.json up to date with the chunk hashes
for some reason. Instead, I'm going the route of generating the same static bundle that would
go to production, and serving it statically with an http server.

The main downside of this is that there's no hot replacement or automatic rebuilding in this
configuration, so it's really meant just to test offline functionality.