testing: Finally, actually fixing the root cause of test flakiness.
The issue here is that when moving an item up in the list, the focused
item may get detached from the DOM (and the same item may or may not get
reattached at the new position). As a result, the item received a blur
event, which is also why we previously needed a hack when moving items
up - this removes that hack too.